[fix] コレクションがなぜだか初期化され続けるバグをfix
