--- Assets/GameLogic/ShopSuperClass.cs
@@ -1,122 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using UnityEngine;
-using SgLibUnite.Singleton;
-
-namespace TeamC
-{
-    /// *MEMO*
-    /// ～セーブデータ～
-    /// NPCの購入数をセーブデータに記憶
-    /// 
-    /// <summary> ShopのSuperクラス </summary>
-    public class ShopSuperClass : SingletonBaseClass<ShopSuperClass>, IShop, IInitializedTarget
-    {
-        [SerializeField, Tooltip("購入可能なNPCのリスト"), Header("The Buyable NPCs")]
-        private List<NPCDataTemplate> npcList;
-
-        /// <summary> NPCの名前に対応したそのNPCの購入数を格納 TName, TBoughtCount </summary>
-        private Dictionary<string, int> _npcShopHistory = new(); // name, bought-count
-
-        protected List<int> _shopHis = new();
-        /// <summary> NPCの購入情報NPC名とその購入数を格納している #sd </summary>
-        protected Dictionary<string, int> GetNPCShopHistory
-        {
-            get { return _npcShopHistory; }
-        }
-
-        public void InitializeObject()
-        {
-            var data = GameObject.FindFirstObjectByType<ClientDataSaverSuperClass>().ReadData();
-            if (data == null)
-            {
-                foreach (var npc in npcList)
-                {
-                    // calculate the bought count link to npc-name
-                    _npcShopHistory.Add(npc.name, 0);
-                }
-            }
-            else
-            {
-                // get each npc level to calculate bought-count
-                var wror = data._saveWarriorLevel - 1;
-                var wzrd = data._saveWizardLevel - 1;
-                var thf = data._saveThiefLevel - 1;
-                var hrmt = data._saveHermitLevel - 1;
-                var pt = data._savePoetLevel - 1;
-
-                _npcShopHistory.Add("Warrior", wror);
-                _npcShopHistory.Add("Wizard", wzrd);
-                _npcShopHistory.Add("Thief", thf);
-                _npcShopHistory.Add("Hermit", hrmt);
-                _npcShopHistory.Add("Poet", pt);
-                
-                _shopHis.Add(wror);
-                _shopHis.Add(wzrd);
-                _shopHis.Add(thf);
-                _shopHis.Add(hrmt);
-                _shopHis.Add(pt);
-            }
-        }
-
-        public void PauseObject()
-        {
-            throw new NotImplementedException();
-        }
-
-        public void ResumeObject()
-        {
-            throw new NotImplementedException();
-        }
-
-        public void FinalizeObject()
-        {
-            // save shop info etc...
-        }
-
-        /// <summary> 購入処理。プレイヤーのリソースを減らす処理と、購入数の＋１処理のみ </summary>
-        public void DecreasePlayerSource(string npcName, decimal cost, Action<int, string> taskToInstantiate)
-        {
-            // get player
-            var player = GameObject.FindFirstObjectByType<PlayerSuperClass>();
-            // process buying 
-            // get bought count
-            var boughtCnt = _npcShopHistory[npcName];
-            player.DecreasePlayerGold(cost);
-            // task to instantiate
-            taskToInstantiate(boughtCnt+1, npcName);
-            // increment bought count 
-            ++_npcShopHistory[npcName];
-        }
-
-        /// <summary> NPC名に応じたNPCの購入数に応じた価格を算出して返す </summary>
-        public decimal CalculateCostToBuy(string npcName)
-        {
-            int boughtCnt = 0;
-            // get bought count
-            // var boughtCnt = _npcShopHistory[npcName];
-            // get target npc
-            boughtCnt = _shopHis[
-                npcName switch { "Warrior" => 0, "Wizard" => 1, "Thief" => 2, "Hermit" => 3, "Poet" => 4 }];
-            var target = npcList.Where(_ => _.Name == npcName).ToList();
-
-            decimal cost = 0;
-            // get base-price
-            if (target.Count == 1) cost = (decimal)target[0].BasePrice;
-            // calculate cost
-            for (int i = 0; i < boughtCnt; i++)
-            {
-                cost *= (decimal)1.15;
-            }
-
-            return cost;
-        }
-
-        protected override void ToDoAtAwakeSingleton()
-        {
-            
-        }
-    }
-}
