--- Assets/GameLogic/ShopSuperClass.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 257fc3c58d42b5440939cb4145e65bc9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
