+++ Assets/Shop/Shop.cs
@@ -1,7 +1,10 @@
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
+using SgLibUnite.Singleton;
 
 namespace TeamC
 {
@@ -18,13 +21,99 @@ namespace TeamC
     #endregion
 
     /// <summary> ショップ機能を提供する </summary>
-    public class Shop : ShopSuperClass
+    public class Shop : SingletonBaseClass<Shop>, IShop, IInitializedTarget
     {
         [SerializeField] private TMP_Text[] _npcLabels;
 
         [SerializeField] private Button[] _shopButton;
 
         private Player _player = new();
+        
+        [SerializeField, Tooltip("購入可能なNPCのリスト"), Header("The Buyable NPCs")]
+        private List<NPCDataTemplate> npcList;
+
+        /// <summary> NPCの名前に対応したそのNPCの購入数を格納 TName, TBoughtCount </summary>
+        protected Dictionary<string, int> _npcShopHistory = new Dictionary<string, int>(); // name, bought-count
+
+        public void InitializeObject()
+        {
+            var data = GameObject.FindFirstObjectByType<ClientDataSaverSuperClass>().ReadData();
+            if (data == null)
+            {
+                foreach (var npc in npcList)
+                {
+                    // calculate the bought count link to npc-name
+                    _npcShopHistory.Add(npc.name, 0);
+                }
+            }
+
+            // get each npc level to calculate bought-count
+            var wror = data._saveWarriorLevel - 1;
+            var wzrd = data._saveWizardLevel - 1;
+            var thf = data._saveThiefLevel - 1;
+            var hrmt = data._saveHermitLevel - 1;
+            var pt = data._savePoetLevel - 1;
+
+            this._npcShopHistory.Add("Warrior", wror);
+            this._npcShopHistory.Add("Wizard", wzrd);
+            this._npcShopHistory.Add("Thief", thf);
+            this._npcShopHistory.Add("Hermit", hrmt);
+            this._npcShopHistory.Add("Poet", pt);
+        }
+
+        public void PauseObject()
+        {
+            throw new NotImplementedException();
+        }
+
+        public void ResumeObject()
+        {
+            throw new NotImplementedException();
+        }
+
+        public void FinalizeObject()
+        {
+            // save shop info etc...
+        }
+
+        /// <summary> 購入処理。プレイヤーのリソースを減らす処理と、購入数の＋１処理のみ </summary>
+        public void DecreasePlayerSource(string npcName, decimal cost, Action<int, string> taskToInstantiate)
+        {
+            // get player
+            var player = GameObject.FindFirstObjectByType<PlayerSuperClass>();
+            // process buying 
+            // get bought count
+            var boughtCnt = this._npcShopHistory[npcName];
+            player.DecreasePlayerGold(cost);
+            // task to instantiate
+            taskToInstantiate(boughtCnt + 1, npcName);
+            // increment bought count 
+            ++this._npcShopHistory[npcName];
+        }
+
+        /// <summary> NPC名に応じたNPCの購入数に応じた価格を算出して返す </summary>
+        public decimal CalculateCostToBuy(string npcName)
+        {
+            // get bought count
+            var boughtCnt = this._npcShopHistory[npcName];
+            // get target npc
+            var target = npcList.Where(_ => _.Name == npcName).ToList();
+
+            decimal cost = 0;
+            // get base-price
+            if (target.Count == 1) cost = (decimal)target[0].BasePrice;
+            // calculate cost
+            for (int i = 0; i < boughtCnt; i++)
+            {
+                cost *= (decimal)1.15;
+            }
+
+            return cost;
+        }
+
+        protected override void ToDoAtAwakeSingleton()
+        {
+        }
 
         void TaskToInstantiateNPC(int boughtCnt, string name)
         {
@@ -55,7 +144,7 @@ namespace TeamC
             {
                 // Superクラスでは、購入数に応じてコストを算出し、
                 // それをプレイヤーへコストの適応をして、購入数を＋１しただけ
-                base.DecreasePlayerSource(name, CalculateNPCCost(name), TaskToInstantiateNPC);
+                DecreasePlayerSource(name, CalculateNPCCost(name), TaskToInstantiateNPC);
                 UpdateButtonDisplayInfo(name);
             }
         }
@@ -67,27 +156,10 @@ namespace TeamC
             //     $"{name} Lv{GetNPCShopHistory[name]} {CalculateNPCCost(name).ToString("F0")}G";
         }
 
-        ///<summary>ボタンが押せるかどうか</summary>
-        /// <param name="value"></param>
-        public void IsPushButton(decimal value)
-        {
-            foreach (Button button in _shopButton)
-            {
-                if (_player.GetCurrentGold() >= CalculateNPCCost(button.name))
-                {
-                    button.interactable = true;
-                }
-                else
-                {
-                    button.interactable = false;
-                }
-            }
-        }
-
         /// <summary> NPCの購入数に応じた価格の算出 </summary>
         public decimal CalculateNPCCost(string npcName)
         {
-            return base.CalculateCostToBuy(npcName);
+            return CalculateCostToBuy(npcName);
         }
 
         private void Start()
@@ -97,7 +169,6 @@ namespace TeamC
 
         private void Update()
         {
-            Debug.Log($"{ _shopHis.Count}");
         }
     }
 }
